+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `Innmind\Immutable\Maybe::memoize()`
+- `Innmind\Immutable\Either::memoize()`
 
 ## 4.12.0 - 2023-03-30